mixins moved to top level folder, renamed BinaryToolMixin to enabledOrDisabledBinaryTool, and made a new mixin: activeOrDisabledBinaryTool.
